enabled add and delete functionalities for Wallets. Added image icon and page for wallet listings. Optimized and clean up functional component template for CRUD pages. Updated field for initial_balance @ wallets table from decimal to float with accuracy.

+++ database/migrations/2021_01_14_010236_create_wallet_table.php
@@ -16,7 +16,7 @@ class CreateWalletTable extends Migration
 		Schema::create('wallets', function (Blueprint $table) {
 			$table->increments('wallet_id');
 			$table->string('wallet_name', 50);
-			$table->integer('initial_balance');
+			$table->float('initial_balance', 10, 2);
 			$table->integer('currency_id')->unsigned();
 			$table->integer('user_id')->unsigned();
 			$table->timestamps();
